+++ crates/stitchy-core/src/files/image.rs
@@ -1,9 +1,6 @@
-
-use crate::{
-    OrderBy, TakeFrom, ImageFilesBuilder, files::FileProperties, image::DynamicImage
-};
+use std::cmp::Ordering;
+use crate::{OrderBy, TakeFrom, ImageFilesBuilder, FileProperties, image::DynamicImage, FileLocation};
 use image::ImageFormat;
-use std::path::Path;
 
 /// A set of image files, storing some file properties internally.
 ///
@@ -14,18 +11,18 @@ use std::path::Path;
 /// Construct using the [`ImageFilesBuilder`] struct.
 ///
 /// See crate-level documentation for examples.
-pub struct ImageFiles {
-    file_list: Vec<FileProperties>
+pub struct ImageFiles<P: FileProperties + Default> {
+    file_list: Vec<P>
 }
 
-impl ImageFiles {
+impl<P: FileProperties + Default> ImageFiles<P> {
 
     /// Create a new [ImageFilesBuilder] for selecting files
-    pub fn builder() -> ImageFilesBuilder {
+    pub fn builder<L: FileLocation<P> + Default>() -> ImageFilesBuilder<P, L> {
         ImageFilesBuilder::default()
     }
 
-    pub(crate) fn new(file_list: Vec<FileProperties>) -> Self {
+    pub(crate) fn new(file_list: Vec<P>) -> Self {
         Self { file_list }
     }
 
@@ -36,17 +33,6 @@ impl ImageFiles {
     /// whole directories.
     pub fn allowed_extensions() -> [&'static str; 5] {
         ["jpg", "jpeg", "png", "gif", "bmp"]
-    }
-
-    /// Mappings of known file extensions to their image format
-    fn extension_formats() -> [(&'static str, ImageFormat); 5] {
-        [
-            (".jpg", ImageFormat::Jpeg),
-            (".jpeg", ImageFormat::Jpeg),
-            (".png", ImageFormat::Png),
-            (".gif", ImageFormat::Gif),
-            (".bmp", ImageFormat::Bmp)
-        ]
     }
 
     /// Get the "main" extension used by a format.
@@ -63,22 +49,6 @@ impl ImageFiles {
         }
     }
 
-    /// Infer the format of an image file based on its file extension
-    pub fn infer_format(file_name: &str) -> Option<ImageFormat> {
-        match file_name.rfind('.') {
-            Some(pos) => {
-                let extension = &file_name[pos..file_name.len()];
-                for &(ext, fmt) in Self::extension_formats().iter() {
-                    if ext == extension {
-                        return Some(fmt);
-                    }
-                }
-                None
-            },
-            None => None
-        }
-    }
-
     /// Get the number of files in the current working set
     pub fn file_count(&self) -> usize {
         self.file_list.len()
@@ -88,7 +58,7 @@ impl ImageFiles {
     pub fn total_size(&self) -> u64 {
         let mut total = 0;
         for file in self.file_list.iter() {
-            total += file.size_bytes;
+            total += file.file_size();
         }
         total
     }
@@ -115,19 +85,24 @@ impl ImageFiles {
         match (order_by, take_from) {
             (OrderBy::Latest, TakeFrom::Start) => {
                 self.file_list.sort_unstable_by(|a, b|
-                    a.modify_time.cmp(&b.modify_time).reverse());
+                    a.modify_time().cmp(&b.modify_time()).reverse());
             },
             (OrderBy::Latest, TakeFrom::End) => {
-                self.file_list.sort_unstable_by(|a, b|
-                    a.modify_time.cmp(&b.modify_time));
+                self.file_list.sort_unstable_by_key(|a| a.modify_time());
             },
             (OrderBy::Alphabetic, TakeFrom::Start) => {
-                self.file_list.sort_unstable_by(|a, b|
-                    a.full_path.cmp(&b.full_path));
+                self.file_list.sort_unstable_by(|a, b| {
+                    let Some(a_path) = a.full_path() else { return Ordering::Equal };
+                    let Some(b_path) = b.full_path() else { return Ordering::Equal };
+                    a_path.cmp(b_path)
+                });
             },
             (OrderBy::Alphabetic, TakeFrom::End) => {
-                self.file_list.sort_unstable_by(|a, b|
-                    a.full_path.cmp(&b.full_path).reverse());
+                self.file_list.sort_unstable_by(|a, b| {
+                    let Some(a_path) = a.full_path() else { return Ordering::Equal };
+                    let Some(b_path) = b.full_path() else { return Ordering::Equal };
+                    a_path.cmp(b_path).reverse()
+                });
             }
         }
         self.file_list.truncate(number_of_files);
@@ -151,22 +126,7 @@ impl ImageFiles {
     pub fn into_image_contents(self, print_info: bool) -> Result<Vec<DynamicImage>, String> {
         let mut images = Vec::with_capacity(self.file_list.len());
         for file in self.file_list {
-
-            let path = Path::new(&file.full_path);
-            let image = image::open(path)
-                .map_err(|_| format!("Failed to open: {:?}", path))?;
-
-            if print_info {
-                if let Some(file_name) = path.file_name() {
-                    let w = image.width();
-                    let h = image.height();
-                    println!(
-                        "Path: {}, w: {}, h: {}, {}",
-                        file_name.to_str().unwrap(),
-                        w, h, crate::util::make_size_string(file.size_bytes));
-                }
-            }
-
+            let image = file.into_image_contents(print_info)?;
             images.push(image);
         }
 
@@ -183,7 +143,7 @@ impl ImageFiles {
             return None;
         }
         let mut all_formats = self.file_list.iter().map(|file_data| {
-            Self::infer_format(&file_data.full_path)
+            file_data.infer_format()
         });
         let first_format = all_formats.next().unwrap();
         match all_formats.all(|fmt| fmt == first_format) {